requirements.txt
google-generativeai
pdfplumber
python-dotenv
python-docx
